refactor(aggregated-assets): split the api per use case (LIVE-35301)
`api.ts` held all five endpoints in one 139-line module, so a change to the
category walk sat beside the paginated search and the id lookup with nothing
separating them.

They now live in three modules named for the question they answer:

- `internals/endpoints/query.ts`      — cursor-paginated search over the aggregate
- `internals/endpoints/lookup.ts`     — "I have ids, hydrate them"
- `internals/endpoints/categories.ts` — "what is in category X"

Each injects into the shared base from `internals/base.ts`, chaining so the
type accumulates every endpoint. `injectEndpoints` mutates and returns that
same object, so this stays one reducer, one middleware and one cache slice —
which `createCurrencyDataSelector` requires, since it hand-scans
`state.assetsDataApi.queries` across every entry.

They sit under `internals/` because chaining means each module exports the
handle the next one builds on, and those are private: each names a partially
built api, and a store configured from one would compile while missing
endpoints. `lint:structure` enforces the same conclusion from the other end — a
barrel may only `export * from "./x"`, so anything not public has to move here
rather than be filtered out in the export.

`api.ts` keeps the public surface it always had: `assetsDataApi` plus the five
hooks. No consumer moves. The two category accessors also leave the public
barrel: they were only ever called by the endpoints that now wrap them, and
nothing outside this package referenced them.
