+++ domain/api/aggregated-assets/src/index.ts
@@ -3,7 +3,6 @@ export * from "./schema";
 export * from "./types";
 export * from "./transforms";
 export * from "./requests";
-export * from "./accessors";
 export * from "./api";
 export * from "./errors";
 export * from "./market";
